Module calc::syntax::visit

Syntax tree traversal.

Each method of the Visitor trait is a hook that can be overridden to customize the behavior when visiting the corresponding type of node. By default, every method recursively visits the substructure of the input by invoking the right visitor method of each of its fields.

Use the walk_*() functions to continue traversing the AST in the default traversal order.

Traits

Visitor
A utility trait for traversing an AST.

Functions

walk_binary_op
Recursively visit a binary operation's left and right operands.

walk_expr
Continue to recursively walk an expression, calling the visitor's visit_atom(), visit_function_call(), or visit_binary_op() method depending on what type of Expr it is.

walk_function_call
Recursively visit each argument in the function call.